init v2

+++ installer/src/install.c
@@ -1,6 +1,7 @@
 #include "common.h"
 #include "util.h"
 #include "screen.h"
+#include "exploit.h"
 #include "install.h"
 
 int create_file(const char *path, mode_t mode, uid_t uid, gid_t gid) {
@@ -12,7 +13,26 @@ int create_file(const char *path, mode_t mode, uid_t uid, gid_t gid) {
     }
 
     int fd = open(path, O_RDWR|O_CREAT);
-    if (fd == -1) return -1;
+    if (fd < 0) return -1;
+    close(fd);
+    sync();
+
+    if (chmod(path, mode) != 0) return -1;
+    if (chown(path, uid, gid) != 0) return -1;
+    sync();
+    return 0;
+}
+
+int write_file(const char *path, void *data, uint32_t size, mode_t mode, uid_t uid, gid_t gid) {
+    if (access(path, F_OK) == 0) {
+        unlink(path);
+        sync();
+    }
+
+    int fd = open(path, O_RDWR|O_CREAT);
+    if (fd < 0) return -1;
+
+    write(fd, data, size);
     close(fd);
     sync();
 
@@ -28,6 +48,40 @@ int set_file_permissions(const char *path, mode_t mode, uid_t uid, gid_t gid) {
     if (chown(path, uid, gid) != 0) return -1;
     sync();
     return 0;
+}
+
+char *replace_substring(const char *str, const char *target, const char *replacement) {
+    if (str == NULL || target == NULL || replacement == NULL) return NULL;
+    size_t str_len = strlen(str);
+    size_t target_len = strlen(target);
+    size_t replacement_len = strlen(replacement);
+
+    size_t occurrences = 0;
+    const char *temp = str;
+    while ((temp = strstr(temp, target))) {
+        temp += target_len;
+        occurrences++;
+    }
+
+    if (occurrences == 0) return strdup(str);
+    size_t new_len = str_len + (occurrences * (replacement_len - target_len)) + 1;
+    char *output = calloc(1, new_len+1);
+
+    const char *current = str;
+    char *dest = output;
+
+    while ((temp = strstr(current, target))) {
+        size_t current_len = temp - current;
+        memcpy(dest, current, current_len);
+        dest += current_len;
+
+        memcpy(dest, replacement, new_len);
+        dest += replacement_len;
+        current = temp + target_len;
+    }
+
+    strcpy(dest, current);
+    return output;
 }
 
 void *load_file(const char *path, size_t *size) {
@@ -105,26 +159,6 @@ int edit_plist(const char *path, void (^action)(CFMutableDictionaryRef plist)) {
     CFRelease(cf_data);
     CFRelease(dict);
     return 0;
-}
-
-int launchctl_unsetenv(void) {
-    char **args = calloc(1, sizeof(char *) * 4);
-    args[0] = "/bin/launchctl";
-    args[1] = "unsetenv";
-    args[2] = "DYLD_INSERT_LIBRARIES";
-    args[3] = NULL;
-
-    pid_t pid = -1;
-    int status = -1;
-    int rv = posix_spawn(&pid, "/bin/launchctl", NULL, NULL, args, NULL);
-    if (rv != 0 || pid == -1) goto done;
-    
-    do { if (waitpid(pid, &status, 0) == -1) goto done; }
-    while (!WIFEXITED(status) && !WIFSIGNALED(status));
-
-done:
-    free(args);
-    return status;
 }
 
 int run_tar(const char *tar_path, const char *output_path) {
@@ -210,24 +244,61 @@ int install_jailbreak(void) {
         access("/Applications/Cydia.app/Cydia", F_OK) == 0 ||
         access("/.cydia_no_stash", F_OK) == 0) no_bootstrap_install = true;
 
-    launchctl_unsetenv();
-    chmod("/private", 0755);
-    chmod("/private/var", 0755);
-
     set_file_permissions("/private/var/mobile", 0755, 501, 501);
     set_file_permissions("/private/var/mobile/Library", 0755, 501, 501);
     set_file_permissions("/private/var/mobile/Library/Preferences", 0755, 501, 501);
     set_file_permissions("/bin/tar", 0755, 0, 0);
     set_file_permissions("/private/var/aquila/aquila", 0755, 0, 0);
     set_file_permissions("/private/var/aquila/_libmis.dylib", 0755, 0, 0);
+    set_file_permissions("/private/var/aquila/amfi_bypass.dylib", 0755, 0, 0);
     set_file_permissions("/private/var/aquila/bootstrap.tar", 0777, 501, 501);
+    set_file_permissions("/private/var/aquila/truststore.tar", 0777, 501, 501);
 
     if (!no_bootstrap_install) {
         set_file_permissions("/private/var/aquila/bootstrap.tar", 0777, 501, 501);
         run_tar("/private/var/aquila/bootstrap.tar", "/");
         print_log("[*] bootstrap installed\n");
+
+        run_tar("/private/var/aquila/truststore.tar", "/");
+        print_log("[*] truststore installed\n");
+
+        if (kinfo->version[0] == 5) {
+            set_file_permissions("/private/var/aquila/safemode5.deb", 0777, 501, 501);
+            set_file_permissions("/private/var/aquila/substrate5.deb", 0777, 501, 501);
+  
+            if (access("/private/var/root/Media/Cydia/AutoInstall", F_OK) != 0) {
+                mkdir("/private/var/root/Media/Cydia/AutoInstall", 0777);
+                chown("/private/var/root/Media/Cydia/AutoInstall", 501, 501);
+            }
+
+            size_t safemode_size = 0;
+            size_t substrate_size = 0;
+            void *safemode_data = load_file("/private/var/aquila/safemode5.deb", &safemode_size);
+            void *substrate_data = load_file("/private/var/aquila/substrate5.deb", &substrate_size);
+
+            if (safemode_data != NULL) {
+                write_file("/private/var/root/Media/Cydia/AutoInstall/safemode5.deb", safemode_data, safemode_size, 0777, 501, 501);
+                munmap(safemode_data, safemode_size);
+            }
+
+            if (substrate_data != NULL) {
+                write_file("/private/var/root/Media/Cydia/AutoInstall/substrate5.deb", substrate_data, substrate_size, 0777, 501, 501);
+                munmap(substrate_data, substrate_size);
+            }
+        }
+
+        FILE *file = fopen("/etc/apt/sources.list.d/aquila.list", "w+");
+        if (file != NULL) {
+            fprintf(file, "deb https://lukezgd.github.io/repo ./\n");
+            fflush(file);
+            fclose(file);
+            sync();
+        }
     } else {
         unlink("/private/var/aquila/bootstrap.tar");
+        unlink("/private/var/aquila/truststore.tar");
+        unlink("/private/var/aquila/safemode5.tar");
+        unlink("/private/var/aquila/substrate5.tar");
     }
 
     clear_mobile_installation_cache();
@@ -261,28 +332,50 @@ int install_jailbreak(void) {
 
     FILE *file = fopen("/etc/launchd.conf", "wb+");
     if (file == NULL) return -1;
-    fprintf(file, "unload /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
-    fprintf(file, "bsexec .. /sbin/mount -u -o rw,suid,dev /\n");
-    fprintf(file, "setenv DYLD_INSERT_LIBRARIES /private/var/aquila/_libmis.dylib\n");
-    fprintf(file, "load /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
-    fprintf(file, "bsexec .. /private/var/aquila/aquila\n");
-    fprintf(file, "unsetenv DYLD_INSERT_LIBRARIES\n");
+    if (kinfo->version[0] == 4) {
+        fprintf(file, "bsexec .. /sbin/mount -u -o rw,suid,dev /\n");
+        fprintf(file, "setenv DYLD_INSERT_LIBRARIES /private/var/aquila/amfi_bypass.dylib\n");
+        fprintf(file, "unload /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
+        fprintf(file, "load /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
+        fprintf(file, "start com.apple.MobileFileIntegrity\n");
+        fprintf(file, "bsexec .. /private/var/aquila/aquila\n");
+        fprintf(file, "unsetenv DYLD_INSERT_LIBRARIES\n");
+    } else {
+        fprintf(file, "unload /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
+        fprintf(file, "bsexec .. /sbin/mount -u -o rw,suid,dev /\n");
+        fprintf(file, "setenv DYLD_INSERT_LIBRARIES /private/var/aquila/amfi_bypass.dylib\n");
+        fprintf(file, "load /System/Library/LaunchDaemons/com.apple.MobileFileIntegrity.plist\n");
+        fprintf(file, "bsexec .. /private/var/aquila/aquila\n");
+        fprintf(file, "unsetenv DYLD_INSERT_LIBRARIES\n");
+    }
 
     fflush(file);
     fclose(file);
     set_file_permissions("/etc/launchd.conf", 0644, 0, 0);
+
+    size_t fstab_size = 0;
+    char *fstab_data = load_file("/etc/fstab", &fstab_size);
+    if (fstab_data == NULL) return -1;
+
+    char *nosuid = replace_substring(fstab_data, ",nosuid,nodev", "");
+    if (nosuid == NULL) return -1;
+
+    char *new_fstab = replace_substring(nosuid, " ro ", " rw ");
+    if (new_fstab == NULL) return -1;
 
     file = fopen("/etc/fstab", "wb+");
     if (file == NULL) return -1;
-    fprintf(file, "/dev/disk0s1s1 / hfs rw 0 1\n");
-    fprintf(file, "/dev/disk0s1s2 /private/var hfs rw 0 2\n");
-
+    fwrite(new_fstab, strlen(new_fstab)+1, 1, file);
     fflush(file);
     fclose(file);
-    set_file_permissions("/etc/launchd.conf", 0644, 0, 0);
 
+    set_file_permissions("/etc/fstab", 0644, 0, 0);
     unlink("/Library/Logs/CrashReporter/Baseband");
     unlink("/private/var/aquila/splashscreen.jp2");
+    unlink("/private/var/aquila/bootstrap.tar");
+    unlink("/private/var/aquila/truststore.tar");
+    unlink("/private/var/aquila/safemode5.tar");
+    unlink("/private/var/aquila/substrate5.tar");
     sync();
     return 0;
 }
